move configuration schema specs
