+++ src/navigate/config/device_schema.py
@@ -34,7 +34,7 @@ from pathlib import Path
 from typing import Optional
 
 # Local Imports
-from navigate.model.devices.configuration_schema import (
+from navigate.config.configuration_schema import (
     CollectionSpec,
     SettingSpec,
     merge_configuration_schemas,
