+++ src/navigate/model/devices/camera/daheng.py
@@ -63,7 +63,7 @@ import numpy as np
 # Local Imports
 from navigate.model.utils.exceptions import UserVisibleException
 from navigate.model.devices.camera.base import CameraBase
-from navigate.model.devices.configuration_schema import SettingSpec
+from navigate.config.configuration_schema import SettingSpec
 from navigate.tools.decorators import log_initialization
 
 # Logger Setup
