+++ src/navigate/model/devices/camera/hamamatsu.py
@@ -38,7 +38,7 @@ from typing import Any, Optional
 
 # Local Imports
 from navigate.model.devices.camera.base import CameraBase
-from navigate.model.devices.configuration_schema import SettingSpec
+from navigate.config.configuration_schema import SettingSpec
 from navigate.model.devices.device_types import SequenceDevice
 from navigate.model.devices.APIs.hamamatsu.HamamatsuAPI import DCAM
 from navigate.tools.decorators import log_initialization
