+++ src/navigate/model/devices/camera/photometrics.py
@@ -42,7 +42,7 @@ from pyvcam.camera import Camera as PyvcamCamera
 
 # Local Imports
 from navigate.model.devices.camera.base import CameraBase
-from navigate.model.devices.configuration_schema import SettingSpec
+from navigate.config.configuration_schema import SettingSpec
 from navigate.model.devices.device_types import IntegratedDevice
 from navigate.tools.decorators import log_initialization
