+++ src/navigate/model/devices/camera/ximea.py
@@ -39,7 +39,7 @@ from ximea import xiapi
 
 # Local Imports
 from navigate.model.devices.camera.base import CameraBase
-from navigate.model.devices.configuration_schema import SettingSpec
+from navigate.config.configuration_schema import SettingSpec
 from navigate.tools.decorators import log_initialization
 
 # Logger Setup
