+++ src/navigate/model/devices/daq/base.py
@@ -39,7 +39,7 @@ from abc import ABC, abstractmethod
 
 # Local Imports
 from navigate.model.waveforms import camera_exposure
-from navigate.model.devices.configuration_schema import SettingSpec
+from navigate.config.configuration_schema import SettingSpec
 from navigate.tools.decorators import log_initialization
 
 # Logger Setup
