+++ src/navigate/model/devices/filter_wheel/ni.py
@@ -43,7 +43,7 @@ from nidaqmx.constants import LineGrouping
 
 # Local Imports
 from navigate.model.devices.filter_wheel.base import FilterWheelBase
-from navigate.model.devices.configuration_schema import CollectionSpec, SettingSpec
+from navigate.config.configuration_schema import CollectionSpec, SettingSpec
 from navigate.model.devices.device_types import NIDevice
 from navigate.tools.decorators import log_initialization
