+++ src/navigate/model/devices/galvo/ni.py
@@ -40,7 +40,7 @@ import nidaqmx
 
 # Local Imports
 from navigate.model.devices.galvo.base import GalvoBase
-from navigate.model.devices.configuration_schema import SettingSpec
+from navigate.config.configuration_schema import SettingSpec
 from navigate.model.devices.device_types import NIDevice
 from navigate.tools.decorators import log_initialization
