+++ src/navigate/model/devices/mirror/imop.py
@@ -39,7 +39,7 @@ import logging
 # Local Imports
 from navigate.model.devices.mirror.base import MirrorBase
 from navigate.model.devices.device_types import IntegratedDevice
-from navigate.model.devices.configuration_schema import SettingSpec
+from navigate.config.configuration_schema import SettingSpec
 from navigate.model.devices.APIs.imagineoptics.imop import IMOP_Mirror
 from navigate.tools.decorators import log_initialization
