+++ src/navigate/model/devices/shutter/asi.py
@@ -38,7 +38,7 @@ from typing import Any
 # Local Imports
 from navigate.model.devices.shutter.base import ShutterBase
 from navigate.model.devices.device_types import SerialDevice
-from navigate.model.devices.configuration_schema import SettingSpec
+from navigate.config.configuration_schema import SettingSpec
 from navigate.tools.decorators import log_initialization
 from navigate.model.devices.APIs.asi.asi_tiger_controller import TigerController
