+++ src/navigate/model/devices/stage/asi.py
@@ -39,7 +39,7 @@ from typing import Any, Optional
 # Local Imports
 from navigate.model.devices.stage.base import StageBase
 from navigate.model.devices.device_types import SerialDevice, IntegratedDevice
-from navigate.model.devices.configuration_schema import SettingSpec
+from navigate.config.configuration_schema import SettingSpec
 from navigate.model.devices.APIs.asi.asi_tiger_controller import (
     TigerController,
     ASIException,
