+++ src/navigate/model/devices/zoom/base.py
@@ -39,7 +39,7 @@ from typing import Any, Optional
 
 # Local Imports
 from navigate.tools.decorators import log_initialization
-from navigate.model.devices.configuration_schema import CollectionSpec, SettingSpec
+from navigate.config.configuration_schema import CollectionSpec, SettingSpec
 
 # Logger Setup
 p = __name__.split(".")[1]
